+++ SHModels/SHModels/Code/Models/Habits/SHActiveDaysDescriptions.h
@@ -15,8 +15,8 @@
 #include "SHDBDueDateConstants.h"
 
 
-char * SH_buildWeekDescription(int32_t weekHash);
-char * SH_buildIntervalLabelDescription_week(int32_t weekHash, int32_t intervalSize);
+char * SH_buildWeekDescription(uint8_t weekHash);
+char * SH_buildIntervalLabelDescription_week(uint8_t weekHash, int32_t intervalSize);
 char * SH_selectIntervalDescription(struct SHActiveDaysValues *activeDays);
 char * (*SH_selectDescriptionBuilderFunc(SHIntervalType intervalType))(int32_t, struct SHActiveDaysValues *);
